git: symbolic-ref
